requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
pinecone
google-generativeai>=0.5.0
cohere>=5.3.0
pypdf>=4.0.0
tqdm>=4.66.0
streamlit>=1.36.0
langchain>=0.2.0
langchain-community>=0.2.0
pinecone-client>=3.0.0
google-generativeai>=0.5.0
cohere>=5.3.0
pypdf>=4.0.0
tqdm>=4.66.0
sentence-transformers